requirements.txt
fastapi
uvicorn
google-cloud-firestore
pandas
pydantic
python-multipart
python-dotenv
pytest
httpx
google-cloud-storage
google-cloud-aiplatform
google-genai
